Small fixes (#16)
* player cant get hurt while being hurt. Also mob ki regen tied to the resource. Introduced a bug on the mobKiBar

* small fixes in the KiBar of the mobs, tweaked some numbers on the resoruce to make the Ki and its regen feel nicer on mobs

* adjusted some mob resources ki value

* player cant get hurt while being hurt. Also mob ki regen tied to the resource. Introduced a bug on the mobKiBar

* small fixes in the KiBar of the mobs, tweaked some numbers on the resoruce to make the Ki and its regen feel nicer on mobs

* adjusted some mob resources ki value

* chore: removed unused code in main.cs

* Fixed start menu hover and set focus with mouse.

+++ Scenes/UI/KIBar/KiBar.cs
@@ -15,18 +15,18 @@ public partial class KiBar : ProgressBar
   {
     if (args[2] is not Player) return;
 
-    if (args[0] is int kiValue && args[1] is int maxKI)
+    if (args[0] is float kiValue && args[1] is float maxKI)
     {
-      MaxValue = maxKI;
-      Value = kiValue;
+      MaxValue = (double) maxKI;
+      Value = (double) kiValue;
     }
   }
   private void OnKiChanged(object sender, object[] args)
   {
     if (args[1] is not Player) return;
-    if (args[0] is int kiValue)
+    if (args[0] is float kiValue)
     {
-      Value = kiValue;
+      Value = (double) kiValue;
     }
   }
